Renamed delete containers to delete instances
