+++ src/classes/Tools/Instances/DeleteInstances.php
@@ -1,9 +1,9 @@
 <?php
-namespace dhope0000\LXDClient\Tools\Containers;
+namespace dhope0000\LXDClient\Tools\Instances;
 
 use dhope0000\LXDClient\Model\Client\LxdClient;
 
-class DeleteContainers
+class DeleteInstances
 {
     public function __construct(LxdClient $lxdClient)
     {
